edit: changeFormat 함수 모듈화
