+++ pages/studyDetail.jsx
@@ -16,6 +16,7 @@ import {
   LOAD_STUDY_MEMBERSHIPS_REQUEST,
 } from '../reducers/study';
 import AddFAB from '../components/AddFAB';
+import { changeFormat } from '../common/changeFormat';
 
 const StyledScreen = styled.div`
   width: calc(100% - 2rem);
@@ -217,29 +218,6 @@ const studyDetail = ({ studyId, token, pk: user }) => {
       Router.pushRoute('/');
     }
   }, []);
-
-  const changeFormat = given => {
-    const day = new Date(given).getDay();
-    const weekday = ['SUN', 'MON', 'TUE', 'WED', 'THU', 'FRI', 'SAT'];
-    const matches = given.split(/(T)|(\+)/g);
-    let hour = matches[3].split(':')[0];
-    const minute = matches[3].split(':')[1];
-    let isAm = true;
-
-    if (hour > 12) {
-      hour -= 12;
-      isAm = false;
-    }
-
-    // 2019-08-17T22:00:00+09:00
-    // 2019. 07. 20 SAT 2:00 PM
-
-    const date = `${matches[0].replace(/-/g, '. ')} ${
-      weekday[day]
-    } ${hour}:${minute} ${isAm ? 'AM' : 'PM'}`;
-
-    return date;
-  };
 
   return (
     <div>
